requirements.txt
# Requirements for DIGITAL LABOUR agents

# LLM Providers
openai>=1.12.0
anthropic>=0.39.0
httpx>=0.27.0

# Core
pydantic>=2.6.0
python-dotenv>=1.0.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
